refactor(i18n): split Messages into a store and a renderer
Move the messages.yml load/merge, the missing-key copy-back, the prefix
lookup, and raw key lookup into a new MessageStore. Messages keeps its
public methods and now pulls raw strings from the store before applying
placeholders, the external resolver, and colors. Behaviour, including
the missing-key warning and bundled-default merge, is unchanged.
